+++ Products/public_server/src/main.rs
@@ -75,3 +75,7 @@ async fn main() -> std::io::Result<()> {
     .run()
     .await
 }
+
+
+#[cfg(test)]
+mod tests;
